scanner: Prepare for integrating remote scanners
Move all code that is only required for local scanners to the new
LocalScanner class. Also change the scan() method in Scanner to take a list
of packages. This allows the Scanner implementation to decide if they are
process in serial order or in parallel.
