+++ scanner/src/main/kotlin/scanners/ScanCode.kt
@@ -21,9 +21,9 @@ package com.here.ort.scanner.scanners
 
 import ch.frankel.slf4k.*
 import ch.qos.logback.classic.Level
+import com.here.ort.scanner.LocalScanner
 
 import com.here.ort.scanner.ScanException
-import com.here.ort.scanner.Scanner
 import com.here.ort.utils.OS
 import com.here.ort.utils.ProcessCapture
 import com.here.ort.utils.getCommandVersion
@@ -33,7 +33,7 @@ import com.here.ort.utils.log
 import java.io.File
 import java.util.regex.Pattern
 
-object ScanCode : Scanner() {
+object ScanCode : LocalScanner() {
     private const val OUTPUT_FORMAT = "json-pp"
     private const val TIMEOUT = 300
